==> 1. BERT Model Prediction.ipynb <==
"""
This script inputs the corpus and
runs several BERT functions to create
several network measures

For more information on BERT - go to https://huggingface.co/transformers/model_doc/bert.html
"""

from transformers import BertTokenizer, BertForMaskedLM, pipeline
import torch
import string
from collections import defaultdict
import nltk
from nltk.corpus import stopwords
import numpy as np
import datetime
import time 
import spacy
import pandas as pd
import re

#Loads the english pipeline from spacy, with english stopwords
nlp = spacy.load("en_core_web_sm")
stop_words = set(stopwords.words('english')) 

import bert_functions as b_funcs

# configure the BERT network 
"""
Two models are available, one is uncased, the other one is cased, 
Change according to the importance of CASE in the sentence
- bert-base-cased: Model is case-sensitive and there is a difference between 'english' and 'English'
- bert-base-uncased: Model is case-insensitive and there is no difference between 'english' and 'English'

"""

# bert_tokenizer = BertTokenizer.from_pretrained('bert-base-cased')
# bert_model = BertForMaskedLM.from_pretrained('bert-base-cased').eval()

"""
Input network which is exported from Gephi with network information
"""

# network = pd.read_csv("MLDA Synonym.csv")
Gephi_output_file_path = "../source_data/gephi_output_cleaned.csv" 

## Link network

"""
Reads exported file from Gephi, converts it into a dataframe
and sets the Label as the index
"""
network = pd.read_csv(Gephi_output_file_path)
network.set_index(network['Label'],inplace=True)

"""
Reads in the entire corpus - input file 
should contain one document per row.

Creates an index column for each document.
"""
### corpus input path to read the original text
corpus_input_csv_path = '../source_data/gme_corpus_inputs_10.csv'

df = pd.read_csv(corpus_input_csv_path).rename(columns = {"snippet":"Text"}) ### load the corpus text 
df = df.reset_index()


def output_predict_metrics_from_masked_words():
    """
    For every document in the corpus, runs
    BERT functions to detect
    masked words and generate network
    metrics for predicted words.

    Returns
    -------
    DataFrame,
        Consists of network measures such as degree centrality, betweenness centrality, modularity class and authority scores.
    """
    res_new = pd.DataFrame()
    for i in range(0, len(df)):
        if i%100==0:
            print(f'{i} / {len(df)} done')
        input_text = df['Text'].loc[i]
        #Creates a dataframe with network measures
        res = b_funcs.key_word_predict_with_network_from_sent(input_text,top_k=5)
        if res is not None:
            res['from_textid'] = df['index'].loc[i]
            res_new = res_new.append(res)
    res_new2 = res_new.reset_index()
    res_new2.to_csv('res.csv', index=False)
    return res_new2
          
result_df = output_predict_metrics_from_masked_words()

result_df.head()
