==> 3. Sentiment Assignment.ipynb <==
"""
This notebook is used to compute sentiment 
distribution score of the predicted emo-denoting words 
based on their distance to the top words represented 
in each modularity class.
"""

import pandas as pd 
import networkx as nx
import numpy as np
from scipy.special import softmax
from collections import defaultdict
from networkx.algorithms.shortest_paths.generic import shortest_path_length as spl

## create undirected network
"""
The following are needed for this notebook:
1) Path to edge_list
2) Path to node_list 
3) Exported gephi output csv file
4) Non-aggregated detected emotion words - pref_with_rfc.csv (from Emo-denoting Prediction without aggregated + corpus)

"""
#Instantiate new Graph object
Graphtype = nx.Graph()

#Path for all input files
edge_list_path = "../source_data/second_iteration_edge_list_updated.csv"
node_list_path = "../source_data/second_iteration_nodes_updated.csv"
gephi_output_csv_path = "../source_data/gephi_output_cleaned.csv"
output_folder="../output/"
emo_pred_file_path = "pred_with_rfc.csv"


## select the top k words from each modularity_class to represent the class
top_k_words = 10

### build the network
edge_list = pd.read_csv(edge_list_path)
node_list = pd.read_csv(node_list_path)
syn_G = nx.from_pandas_edgelist(edge_list, edge_attr='weight', create_using=Graphtype)

## read gephi output graph for computed distance, centrality, authority values
network = pd.read_csv(gephi_output_csv_path)
network.reset_index()
network.set_index(network['Id'],inplace=True)
mapping =network['Label'].to_dict()
G = nx.relabel_nodes(syn_G, mapping, copy=False)

## Sorts network by Authority scores and groups by modularity class
selected_top = network.sort_values('Authority',ascending=False).groupby('modularity_class').head(top_k_words)
## generate dic with cluster number as key and representing words list as the value
selected_top_dic = selected_top.groupby('modularity_class')['Label'].agg(list).to_dict()
### import the emo_pred_output from Emo-denoting prediction notebook
pred = pd.read_csv('pred_with_rfc.csv')


def emo_distribution_cal(top_cluster_word_dict = selected_top_dic,source=None,network_Graph=G,row=None):
    """
    Parameters
    ----------
    top_cluster_word_dict: dictionary,
        Dictionary of the top k words for each modularity class
    
    source: str,
        The source word - default is None
        
    network_Graph: networkX Graph,
        NetworkX Graph object
    
    row: 
        
    output softmax function value for the probability of the emotion based on the score 
    """
    emo_score_dist = []
    if row['emo?']:
        source=row['word'].split('_')[0]
    else:
        return 
    for key, top_words in top_cluster_word_dict.items():
        total_distance = 0
        count =0
        ### loop for each top word in one cluster
        for top_word in top_words:
            try: 
                """
                spl retrieves the shortest distance from the source to the top_word in the network
                """
                total_distance += spl(network_Graph,source,top_word)
                count+=1
            except:
                """
                when either source or top_word in the one cluster not found in the network 
                or they are not reachable => just pass 
                """
                pass
        ### shorter distance between nodes => closer the relationship => 1/ avg_distance of the cluster
        if count !=0:
            avg_distance = total_distance/count
            emo_score_dist.append(1/avg_distance) 
        else:
            emo_score_dist.append(0)
        
    return softmax(emo_score_dist)


### emotion score distribution of all the clusters
total_emo_res = pred.apply(lambda x: emo_distribution_cal(row=x),axis=1)
pred["emo_dist_prob"] = total_emo_res

def emo_cluster_assignment(row):
    if row['emo?']==1:
        return np.array(row['emo_dist_prob']).argsort()[::-1]
    else:
        return 0

pred['emo_dist_cluster_order']=pred.apply(lambda x: emo_cluster_assignment(x),axis=1)


### output results
pred.to_csv('emo_assignment.csv',index=False)

emo_df = pd.read_csv('emo_assignment.csv')

emo_subset = emo_df[['emo?', 'word', 'emo_dist_prob', 'emo_dist_cluster_order']]

main_df = pd.read_csv('res.csv')
main_df.head()

main_df.shape, emo_subset.shape

combined_df = main_df.merge(emo_subset, how='left', left_on='index', right_on='word').drop_duplicates()

combined_df.shape

combined_df['emo_dist_cluster_order'] = combined_df['emo_dist_cluster_order'].fillna(-1)
combined_df['emo_dist_prob'] = combined_df['emo_dist_prob'].fillna(-1)

combined_df['emo_dist_cluster_order'] = combined_df['emo_dist_cluster_order'].replace('0', -1)

cleaned_df = combined_df.dropna()

cleaned_df.head()

cleaned_df['emo_dist_cluster_order'] = cleaned_df['emo_dist_cluster_order'].to_list()

cleaned_df['emo_dist_cluster_order'].dtype

testing = cleaned_df.iloc[1][17]

def list_mapping(string):
    res = list(map(float, string[1:-1].split()))
    return res

cleaned_df['emo_dist_cluster_order'] = cleaned_df['emo_dist_cluster_order'].apply(lambda x: list_mapping(x) if x != -1 else -1)

cleaned_df['emo_dist_prob'] = cleaned_df['emo_dist_prob'].apply(lambda x: list_mapping(x) if x != -1 else -1)

cleaned_df['emo_dist_prob'] = cleaned_df['emo_dist_prob'].to_list()

cleaned_df['emo_dist_cluster_order'] = cleaned_df['emo_dist_cluster_order'].to_list()

cleaned_df.head()

exploded_df = cleaned_df.apply(pd.Series.explode)

exploded_df.to_csv('emo_exploded_df.csv', index=False)

exploded_df.head(25)

#final_df = combined_df[~combined_df['emo_dist_prob'].isnull()]

#print(final_df.shape)
#final_df.head(10)

#emo_dist_prob - explode two columns and export to bigquery - convert arrays to list and explode them
#naming conventions, import exports, SQL, tableau
#reddit tags (no emotions = -1)



final_df['from_textid'].unique()
