==> 4. BigQuery Export.ipynb <==
import pandas as pd

exploded_df = pd.read_csv('emo_exploded_df.csv')
exploded_df.head()

!pip install google-api-python-client

import connect_cloud
import connect_google_sheets

exploded_df.head(15) #drop predicted scores, prediction(2nd column), content_stonks

final_df = exploded_df.drop(columns=['prediction', 'pred_auth', 'pred_deg', 'pred_class'])

columns_list = list(final_df.columns)
columns_list[10] = 'emo'


final_df.columns = columns_list

final_df.to_csv('final_emo_df.csv', index=False)

connect_cloud.writeToBigQuery(final_df, 'content_stonks', 'tableau_df_analysis', mode='replace')

final_df.shape
